refactored filter function as component
