+++ src/Components/Dashboard.jsx
@@ -2,7 +2,7 @@
 import { useState } from "react";
 //Project files
 import bursaries from "Data/bursaries.json";
-import BursaryItem from "./BursaryItem";
+import FilterBursaries from "./FilterBursaries";
 import { useFilter } from "state/FilterContext";
 import SelectCategory from "./SelectCategory";
 import SelectMonth from "./SelectMonth";
@@ -25,41 +25,12 @@ export default function Dashboard() {
     />
   );
 
-  function userClickedFilter(a, b) {
-    const byMonth = bursaries.filter((item) => item.dueDate === `${month}`);
-    const byCategory = bursaries.filter(
-      (item) => item.category === `${academic_category}`
-    );
-    if (a === "All" && b === "All") {
-      return bursaries.map((item) => <BursaryItem key={item.id} item={item} />);
-    } else if (a === "All" && b !== "All") {
-      const filteredList = byCategory;
-      return filteredList.map((item) => (
-        <BursaryItem key={item.id} item={item} />
-      ));
-    } else if (a !== "All" && b === "All") {
-      const filteredList = byMonth;
-      return filteredList.map((item) => (
-        <BursaryItem key={item.id} item={item} />
-      ));
-    } else {
-      const filteredList = byMonth.filter(
-        (item) => item.category === `${academic_category}`
-      );
-      return filteredList.map((item) => (
-        <BursaryItem key={item.id} item={item} />
-      ));
-    }
-  }
-
-  const bursary_list = userClickedFilter(month, academic_category);
-
   return (
     <div>
       <SelectCategory />
       <SelectMonth />
       {searchbox}
-      {bursary_list}
+      <FilterBursaries />
     </div>
   );
 }
